Update decrypt.go
Remove else statement. Because the encrypted assertion dose not come with the either X509IssuerSerial or X509Certificate within the SAML response. So, the else statement will return unexpected error by mistake. Since it is already state that "verification is not security-critical." in comments above, the else statement will be removed.